# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Joe Kuntani Platform API
  version: 0.1.0
  description: Versioned contract for the standalone Joe Kuntani platform.
  license:
    name: Proprietary - all rights reserved
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Relative API base; clients configure the environment origin.
security: []
...
paths:
...
  /api/admin/crm/enquiries/{id}/workflow:
    get:
      operationId: getCRMEnquiryWorkflow
      summary: Read internal notes, tasks, immutable stage history and proposal attachments
      tags: [CRM]
      security: [{ staffSession: [] }]
      parameters:
        [
          {
            name: id,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          },
        ]
      responses:
        "200":
          {
            description: Private CRM workflow,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/CRMWorkflow" } },
              },
          }
        "4XX": { $ref: "#/components/responses/Problem" }
  /api/admin/crm/enquiries/{id}/notes:
    post:
      operationId: addCRMEnquiryNote
      summary: Add an audited internal enquiry note
      tags: [CRM]
      security: [{ staffSession: [], csrfToken: [] }]
      parameters:
        [
          {
            name: id,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          },
        ]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/CRMNoteInput" } },
            },
        }
      responses:
        "201":
          {
            description: Note created,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/CRMNote" } },
              },
          }
        "4XX": { $ref: "#/components/responses/Problem" }
  /api/admin/crm/enquiries/{id}/tasks:
    post:
      operationId: addCRMEnquiryTask
      summary: Add an assigned prioritized follow-up task and queue an internal notification
      tags: [CRM]
      security: [{ staffSession: [], csrfToken: [] }]
      parameters:
        [
          {
            name: id,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          },
        ]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/CRMTaskInput" } },
            },
        }
      responses:
        "201":
          {
            description: Task created,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/CRMTask" } },
              },
          }
        "4XX": { $ref: "#/components/responses/Problem" }
  /api/admin/crm/enquiries/{id}/tasks/{taskID}/complete:
    post:
      operationId: completeCRMEnquiryTask
      summary: Complete an open follow-up task exactly once
      tags: [CRM]
      security: [{ staffSession: [], csrfToken: [] }]
      parameters:
        - {
            name: id,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
        - {
            name: taskID,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
      responses:
        "200":
          {
            description: Completed task,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/CRMTask" } },
              },
          }
        "4XX": { $ref: "#/components/responses/Problem" }
  /api/admin/crm/enquiries/{id}/attachments:
    post:
      operationId: addCRMProposalAttachment
      summary: Reference a protected ready proposal document
      tags: [CRM]
      security: [{ staffSession: [], csrfToken: [] }]
      parameters:
        [
          {
            name: id,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          },
        ]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/CRMAttachmentInput" } },
            },
        }
      responses:
        "201":
          {
            description: Protected attachment reference created,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/CRMAttachment" } },
              },
          }
        "4XX": { $ref: "#/components/responses/Problem" }
  /api/admin/crm/enquiries/{id}/attachments/{attachmentID}/access:
    post:
      operationId: accessCRMProposalAttachment
      summary: Audit and issue a five-minute private proposal download URL
      tags: [CRM]
      security: [{ staffSession: [], csrfToken: [] }]
      parameters:
        - {
            name: id,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
        - {
            name: attachmentID,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
      responses:
        "200":
          {
            description: Expiring private access,
            content:
              {
                application/json:
                  {
                    schema:
                      { $ref: "#/components/schemas/CRMAttachmentAccess" },
                  },
              },
          }
        "4XX": { $ref: "#/components/responses/Problem" }
...
  /api/admin/crm/proposal-download:
    get:
      operationId: downloadCRMProposalAttachment
      summary: Resolve an authenticated short-lived proposal attachment URL
      tags: [CRM]
      security: [{ staffSession: [] }]
      parameters:
        - {
            name: asset,
            in: query,
            required: true,
            schema: { type: string, format: uuid },
          }
        - {
            name: expires,
            in: query,
            required: true,
            schema: { type: integer, format: int64 },
          }
        - {
            name: signature,
            in: query,
            required: true,
            schema: { type: string, minLength: 43, maxLength: 64 },
          }
      responses:
        "200":
          {
            description: Proxied private proposal document,
            content:
              {
                application/pdf: { schema: { type: string, format: binary } },
                application/vnd.openxmlformats-officedocument.wordprocessingml.document:
                  { schema: { type: string, format: binary } },
              },
          }
        "403": { $ref: "#/components/responses/Problem" }
  /api/public/ticket-orders:
    post:
      operationId: createPublicTicketOrder
      summary: Create an idempotent pending order with a short atomic inventory hold
      tags: [Ticketing]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 16, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateTicketOrderRequest" }
      responses:
        "201":
          description: Pending order and inventory hold created, or the original receipt replayed
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TicketOrderReceipt" }
        "400": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "415": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /api/public/ticket-orders/{reference}/checkout:
    post:
      operationId: createTicketCheckout
      summary: Create or replay a provider-neutral checkout session for an active held order
      tags: [Ticketing]
      parameters:
        - name: reference
          in: path
          required: true
          schema: { type: string, pattern: "^JKT-[0-9]{4}-[A-Z0-9]{8}$" }
        - name: Order-Access-Key
          in: header
          required: true
          schema: { type: string, minLength: 16, maxLength: 128 }
      responses:
        "200":
          description: Credential-free HTTPS payment session
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CheckoutSession" }
        "401": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /api/public/ticket-orders/{reference}:
    get:
      operationId: getSecureTicketConfirmation
      summary: Read paid tickets using a possession-bound expiring access token
      tags: [Ticketing]
      parameters:
        - name: reference
          in: path
          required: true
          schema: { type: string, pattern: "^JKT-[0-9]{4}-[A-Z0-9]{8}$" }
        - name: Order-Access-Key
          in: header
          required: true
          schema: { type: string, minLength: 16, maxLength: 512 }
      responses:
        "200":
          description: Masked order and one signed bearer per paid admission
          headers:
            Cache-Control:
              schema: { type: string, const: "private, no-store" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TicketConfirmation" }
        "401": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /api/webhooks/payments/{provider}:
    post:
      operationId: receivePaymentWebhook
      summary: Verify and idempotently apply a raw provider-signed payment event
      tags: [Ticketing]
      parameters:
        - name: provider
          in: path
          required: true
          schema: { type: string, minLength: 1, maxLength: 100 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "204": { description: Verified event processed or already processed }
        "401": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /api/public/enquiries:
    post:
      operationId: submitPublicEnquiry
      summary: Submit an enquiry and atomically queue acknowledgement and internal alert
      tags: [Enquiries]
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 16, maxLength: 128 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EnquirySubmission" }
      responses:
        "201":
          description: Stored once; replaying the idempotency key returns the same reference
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnquiryReceipt" }
        "400": { $ref: "#/components/responses/Problem" }
        "415": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /api/public/enquiries/challenge:
    get:
      operationId: getPublicEnquiryChallenge
      summary: Read the browser CAPTCHA configuration for public enquiries
      tags: [Enquiries]
      responses:
        "200":
          description: CAPTCHA challenge configuration; the site key is public
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EnquiryChallenge" }
        "405": { $ref: "#/components/responses/Problem" }
  /api/public/services:
    get:
      operationId: listPublicServices
      summary: List active approved services in display order
      tags: [Services]
      responses:
        "200":
          description: Active services ordered by display position
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServiceList" }
        "400": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
  /api/public/services/{slug}:
    get:
      operationId: getPublicService
      summary: Read one active approved service by immutable slug
      tags: [Services]
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            {
              type: string,
              pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$",
              maxLength: 120,
            }
      responses:
        "200":
          description: Active service
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Service" }
        "404": { $ref: "#/components/responses/Problem" }
  /api/admin/services:
    get:
      operationId: listAdminServices
      summary: List active and inactive services for content management
      tags: [Services]
      security: [{ staffSession: [] }]
      responses:
        "200":
          description: All services ordered by display position
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ServiceList" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
    post:
      operationId: createService
      summary: Create and audit a service with immutable external identifiers
      tags: [Services]
      security: [{ staffSession: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ServiceInput" }
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Service" }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /api/admin/services/order:
    put:
      operationId: reorderServices
      summary: Atomically replace and audit the complete service display order
      tags: [Services]
      security: [{ staffSession: [], csrfToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [items]
              properties:
                items:
                  {
                    type: array,
                    minItems: 1,
                    maxItems: 100,
                    items:
                      {
                        type: object,
                        additionalProperties: false,
                        required: [id, version],
                        properties:
                          {
                            id: { type: string, format: uuid },
                            version:
                              { type: integer, format: int64, minimum: 1 },
                          },
                      },
                  }
      responses:
        "204": { description: Complete display order replaced }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
...
  /api/admin/services/{serviceID}/active:
    patch:
      operationId: setServiceActive
      summary: Publish or unpublish a service and audit the lifecycle transition
      tags: [Services]
      security: [{ staffSession: [], csrfToken: [] }]
      parameters:
        - name: serviceID
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [active, version]
              properties:
                active: { type: boolean }
                version: { type: integer, format: int64, minimum: 1 }
      responses:
        "204": { description: Service lifecycle state updated }
        "400": { $ref: "#/components/responses/Problem" }
        "401": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "404": { $ref: "#/components/responses/Problem" }
  /api/public/settings:
    get:
      operationId: getPublicSettings
      summary: Read the latest approved, published and public-safe global settings
      tags: [Settings]
      responses:
        "200":
          description: Published settings; integration and team fields are structurally excluded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicSettings" }
        "404": { $ref: "#/components/responses/Problem" }
...
  /api/public/events:
    get:
      operationId: listPublicEvents
      summary: List published events with live ticket availability
      tags: [Events]
      responses:
        "200":
          description: Published events ordered by start time
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicEventList" }
        "4XX": { $ref: "#/components/responses/Problem" }
        "503": { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    staffSession:
      type: apiKey
      in: cookie
      name: jk_admin_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  responses:
    Problem:
      description: Request rejected
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    CRMNoteInput:
      type: object
      additionalProperties: false
      required: [body]
      properties: { body: { type: string, minLength: 1, maxLength: 4000 } }
...
    CRMAttachmentInput:
      type: object
      additionalProperties: false
      required: [asset_id, label]
      properties:
        asset_id: { type: string, format: uuid }
        label: { type: string, minLength: 1, maxLength: 200 }
...
    CRMWorkflow:
      type: object
      additionalProperties: false
      required: [notes, tasks, stage_history, attachments]
      properties:
        notes: { type: array, items: { $ref: "#/components/schemas/CRMNote" } }
        tasks: { type: array, items: { $ref: "#/components/schemas/CRMTask" } }
        stage_history:
          {
            type: array,
            items: { $ref: "#/components/schemas/CRMStageHistory" },
          }
        attachments:
          { type: array, items: { $ref: "#/components/schemas/CRMAttachment" } }
...